# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ArticleOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "source_slug": {
            "title": "Source Slug",
            "type": "string"
          },
          "story_cluster_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Cluster Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "description": "Canonical publisher URL. Always link out to this.",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "snippet",
          "image_url",
          "url",
          "language",
          "published_at",
          "source_name",
          "source_slug",
          "story_cluster_id"
        ],
        "title": "ArticleOut",
        "type": "object"
      },
...
      "FollowIn": {
        "properties": {
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "topic_id"
        ],
        "title": "FollowIn",
        "type": "object"
      },
      "FollowOut": {
        "properties": {
          "followed_at": {
            "format": "date-time",
            "title": "Followed At",
            "type": "string"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "topic_id",
          "followed_at"
        ],
        "title": "FollowOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeExportOut": {
        "properties": {
          "follows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Follows",
            "type": "array"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "History",
            "type": "array"
          },
          "profile": {
            "additionalProperties": true,
            "title": "Profile",
            "type": "object"
          },
          "saves": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Saves",
            "type": "array"
          }
        },
        "required": [
          "profile",
          "saves",
          "follows",
          "history"
        ],
        "title": "MeExportOut",
        "type": "object"
      },
      "MeOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "has_beta_access": {
            "title": "Has Beta Access",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "preferred_languages": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Languages",
            "type": "array"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "preferred_languages",
          "role",
          "has_beta_access",
          "created_at"
        ],
        "title": "MeOut",
        "type": "object"
      },
      "MeUpdateIn": {
        "properties": {
          "preferred_languages": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Preferred Languages",
            "type": "array"
          }
        },
        "required": [
          "preferred_languages"
        ],
        "title": "MeUpdateIn",
        "type": "object"
      },
...
      "SaveIn": {
        "properties": {
          "article_id": {
            "title": "Article Id",
            "type": "integer"
          }
        },
        "required": [
          "article_id"
        ],
        "title": "SaveIn",
        "type": "object"
      },
      "SaveOut": {
        "properties": {
          "article": {
            "$ref": "#/components/schemas/ArticleOut"
          },
          "saved_at": {
            "format": "date-time",
            "title": "Saved At",
            "type": "string"
          }
        },
        "required": [
          "article",
          "saved_at"
        ],
        "title": "SaveOut",
        "type": "object"
      },
      "SavePageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SaveOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "SavePageOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Personalised, multilingual news.",
    "title": "JustNews API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/follows": {
      "get": {
        "operationId": "list_follows_v1_follows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FollowOut"
                  },
                  "title": "Response List Follows V1 Follows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Follows",
        "tags": [
          "follows"
        ]
      },
      "post": {
        "operationId": "create_follow_v1_follows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Follow",
        "tags": [
          "follows"
        ]
      }
    },
    "/v1/follows/{topic_id}": {
      "delete": {
        "operationId": "delete_follow_v1_follows__topic_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "topic_id",
            "required": true,
            "schema": {
              "title": "Topic Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Follow",
        "tags": [
          "follows"
        ]
      }
    },
...
    "/v1/me": {
      "delete": {
        "operationId": "delete_me_v1_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Me",
        "tags": [
          "me"
        ]
      },
      "get": {
        "operationId": "get_me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "update_me_v1_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/export": {
      "get": {
        "operationId": "export_me_v1_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeExportOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Me",
        "tags": [
          "me"
        ]
      }
    },
...
    "/v1/saves": {
      "get": {
        "operationId": "list_saves_v1_saves_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavePageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Saves",
        "tags": [
          "saves"
        ]
      },
      "post": {
        "operationId": "create_save_v1_saves_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Save",
        "tags": [
          "saves"
        ]
      }
    },
    "/v1/saves/{article_id}": {
      "delete": {
        "operationId": "delete_save_v1_saves__article_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "title": "Article Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Save",
        "tags": [
          "saves"
        ]
      }
    },
...
  }
}